requirements_api.txt
fastapi==0.109.0
uvicorn==0.27.0
snowflake-connector-python==3.6.0
pydantic==2.6.0